Moved more sql to database from account.
